Bugfix: Use correct expression to disable timer
Due to a wrongly crafted bitmasking expression, timer function
timer_stop_and_reset() did reset the timer without stopping it. This led
to intermittent erratic PTT control that was very challenging to track
down. Pascal HB9HCG was dedicated enough for us to finally track it
down and fix it.

The failure mode was as follows: After clearing of PTT, the timeout
timer was reset but not stopped. In consequence, that timer would exceed
the timeout period in the absence of further PTT assertions within the
timeout period. A subsequent PTT assertion past that timer expiry would
momentarily assert PTT and clear it shortly thereafter due to TOT
action. Further assertions of PTT would reinstate correct logic of
operation, thus avoiding catastrophic and permanent failure of PTT
control.